+++ src/components.d.ts
@@ -560,6 +560,18 @@ export namespace Components {
     'productLinkFormat'?: string;
     'skeleton'?: boolean;
   }
+  interface ManifoldServiceCardView {
+    'asCard'?: boolean;
+    'description'?: string;
+    'hideTags'?: boolean;
+    'isFeatured'?: boolean;
+    'isFree'?: boolean;
+    'label'?: string;
+    'loading'?: boolean;
+    'logo'?: string;
+    'name'?: string;
+    'productId'?: string;
+  }
   interface ManifoldServiceView {
     'asCard'?: boolean;
     'description'?: string;
@@ -908,6 +920,12 @@ declare global {
     new (): HTMLManifoldServiceCardElement;
   };
 
+  interface HTMLManifoldServiceCardViewElement extends Components.ManifoldServiceCardView, HTMLStencilElement {}
+  var HTMLManifoldServiceCardViewElement: {
+    prototype: HTMLManifoldServiceCardViewElement;
+    new (): HTMLManifoldServiceCardViewElement;
+  };
+
   interface HTMLManifoldServiceViewElement extends Components.ManifoldServiceView, HTMLStencilElement {}
   var HTMLManifoldServiceViewElement: {
     prototype: HTMLManifoldServiceViewElement;
@@ -1000,6 +1018,7 @@ declare global {
     'manifold-resource-status-view': HTMLManifoldResourceStatusViewElement;
     'manifold-select': HTMLManifoldSelectElement;
     'manifold-service-card': HTMLManifoldServiceCardElement;
+    'manifold-service-card-view': HTMLManifoldServiceCardViewElement;
     'manifold-service-view': HTMLManifoldServiceViewElement;
     'manifold-skeleton-img': HTMLManifoldSkeletonImgElement;
     'manifold-skeleton-text': HTMLManifoldSkeletonTextElement;
@@ -1581,6 +1600,18 @@ declare namespace LocalJSX {
     'productLinkFormat'?: string;
     'skeleton'?: boolean;
   }
+  interface ManifoldServiceCardView extends JSXBase.HTMLAttributes<HTMLManifoldServiceCardViewElement> {
+    'asCard'?: boolean;
+    'description'?: string;
+    'hideTags'?: boolean;
+    'isFeatured'?: boolean;
+    'isFree'?: boolean;
+    'label'?: string;
+    'loading'?: boolean;
+    'logo'?: string;
+    'name'?: string;
+    'productId'?: string;
+  }
   interface ManifoldServiceView extends JSXBase.HTMLAttributes<HTMLManifoldServiceViewElement> {
     'asCard'?: boolean;
     'description'?: string;
@@ -1678,6 +1709,7 @@ declare namespace LocalJSX {
     'manifold-resource-status-view': ManifoldResourceStatusView;
     'manifold-select': ManifoldSelect;
     'manifold-service-card': ManifoldServiceCard;
+    'manifold-service-card-view': ManifoldServiceCardView;
     'manifold-service-view': ManifoldServiceView;
     'manifold-skeleton-img': ManifoldSkeletonImg;
     'manifold-skeleton-text': ManifoldSkeletonText;