+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fixed the provision button requiring the label to be set, preventing or automatic label generation from working.
+- Fixed name for `manifold-service-view` to be `manifold-service-card-view` to match documentation.
 
 ## [0.4.3]
 ### Fixed